+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.3
+* #2415: add RequestID to error object.
+* #2467: replace IndexedDBShim by WebSQL adapter.
+
 ## 1.0.2 (2013-07-10)
 * Bugfix: IndexedDBShim was broken due to "use strict", downgraded version.
 * Bugfix: Network requests on Titanium were instantiated incorrectly.